refactor: change file name from kebab-case to camelCase
