+++ src/serializers/cars/createCarsSerializer.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
 import type { Cars } from '../../../generated/prisma/index.js';
-import type { CreateCarResponseDto } from '../../dtos/cars/create-cars-response.dto.js';
+import type { CreateCarResponseDto } from '../../dtos/cars/createCars-responseDto.js';
 
 /**
  * Cars 객체와 관련 carModel(제조사, 모델명, 타입) 정보를 받아
